requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
typing
bcrypt==4.0.1
cryptography
fastapi
psycopg2-binary
python-jose
python-multipart
SQLAlchemy
uvicorn
pytest
httpx
jinja2